Format updates
Created format updates to be able to read our stock outcomes easier. With green for positive and red for negatives, we can determine which stocks did better more quickly. A button was also created for easy access use for those who receive this data on their end, and can manage the data they're trying to populate with ease. One button to populate the data in general, one to populate it with all the format changes, and the other to clear the data.